\todo Pobranie z tokena nazwy uzytkownika i potem pobranie jego id

+++ backend/src/main/java/pbs/edu/rekrutacja/services/FileService.java
@@ -10,10 +10,10 @@ import java.util.Optional;
 public interface FileService {
     Page<File> getFiles(Pageable pageable);
 
-    Optional<File> getFile(Integer fileId);
+    Optional<File> getFile(Long fileId);
 
     File setFile(File file);
 
     @Transactional
-    void deleteFile(Integer fileId);
+    void deleteFile(Long fileId);
 }
